412 - Make timestamp generator monotonic
* Monotonic timestamp tests now use a duration instead of an arbitrary
  number of generated timestamps.
* Fix compiler errors for the 64-bit format specifiers.

+++ src/timestamp_generator.cpp
@@ -80,7 +80,9 @@ int64_t MonotonicTimestampGenerator::compute_next(int64_t last) {
                  "microseconds behind the last generated timestamp (%lld). "
                  "The next generated timestamp will be artificially incremented "
                  "to guarantee montonicity.",
-                 current, last - current, last);
+                 static_cast<long long>(current),
+                 static_cast<long long>(last - current),
+                 static_cast<long long>(last));
       }
     }
